+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added
-- add to `printAllObjects()` optional variadic parameter `entityName`.
+- add to `NSManagedObjectContext.printAllObjects()` optional variadic parameter `entityName`.
+- add `NSManagedObjectContext.deleteAllObjects()`.
 
 ## [1.1.5] - 2021-05-15